menopause_qol/requirements.txt
# Data Manipulation & Analysis
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning & AI
scikit-learn>=1.3.0
joblib>=1.3.0
shap>=0.42.0

# Statistical Analysis
pingouin>=0.5.3
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
ydata-profiling>=4.5.0

# Reporting
fpdf2>=2.7.0